+++ ant_design_qt/packages/ant_design_qt/src/widgets/button_style.cpp
@@ -80,7 +80,7 @@ QColor withAlpha(const QColor& color, double alpha) {
 bool isStableChannel(int value) { return value >= 0 && value <= 255; }
 
 QColor resolveAlphaColor(const QColor& frontColor, const QColor& backgroundColor) {
-  if (frontColor.alphaF() < 1.0) {
+  if (static_cast<double>(frontColor.alphaF()) < 1.0) {
     return frontColor;
   }
