+++ snow_shot/src/presentation/ocr/screenshotimageconversionview.cpp
@@ -123,7 +123,8 @@ void ScreenshotImageConversionView::render() {
     const QColor fill = theme.colorFillTertiary;
     const QColor base = theme.colorBgContainer;
     const auto blended = [fill](int foreground, int background) {
-        return qRound(fill.alphaF() * foreground + (1.0 - fill.alphaF()) * background);
+        return qRound(static_cast<double>(fill.alphaF()) * foreground +
+                      (1.0 - static_cast<double>(fill.alphaF())) * background);
     };
     const QColor codeBackground(blended(fill.red(), base.red()),
                                 blended(fill.green(), base.green()),
