+++ snow_shot/src/presentation/ocr/screenshotocrrecognitionservice.cpp
@@ -149,8 +149,9 @@ QPolygonF quadFromValues(const float* points, const QRectF& canvasRect, const QS
     QPolygonF polygon;
     polygon.reserve(4);
     for (int index = 0; index < 4; ++index) {
-        polygon.push_back(QPointF(canvasRect.left() + points[index * 2] * scaleX,
-                                  canvasRect.top() + points[index * 2 + 1] * scaleY));
+        polygon.push_back(
+            QPointF(canvasRect.left() + static_cast<qreal>(points[index * 2]) * scaleX,
+                    canvasRect.top() + static_cast<qreal>(points[index * 2 + 1]) * scaleY));
     }
     return polygon;
 }
@@ -283,7 +284,8 @@ class ScreenshotOcrTransport final : public QObject {
         uchar* header = m_mapping + slot * m_slotBytes;
         writeU32(header + kSlotStateOffset, kSlotFree);
         for (int row = 0; row < image.height(); ++row)
-            std::memcpy(header + kSlotHeaderBytes + row * stride, image.constScanLine(row), stride);
+            std::memcpy(header + kSlotHeaderBytes + row * stride, image.constScanLine(row),
+                        static_cast<std::size_t>(stride));
         writeU64(header + kSlotSequenceOffset, sequence);
         writeU32(header + kSlotWidthOffset, static_cast<quint32>(image.width()));
         writeU32(header + kSlotHeightOffset, static_cast<quint32>(image.height()));
@@ -342,7 +344,7 @@ class ScreenshotOcrRecognitionService::Impl final {
           m_proxyUrl(options.proxyUrl), m_modelType(options.modelType),
           m_backendPreference(preference) {
         m_queueClock.start();
-        m_slots.resize(m_workerLimit);
+        m_slots.resize(static_cast<std::size_t>(m_workerLimit));
         m_transportThread.setObjectName(QStringLiteral("snow-ocr-transport"));
         m_localPool.setMaxThreadCount(m_workerLimit);
         if (!options.processPath.trimmed().isEmpty() &&
@@ -539,7 +541,7 @@ class ScreenshotOcrRecognitionService::Impl final {
                 m_ready = false;
                 m_processStopReason = ProcessStopReason::Cancelled;
                 m_jobs.erase(it);
-                m_slots[job->slot].reset();
+                m_slots[static_cast<std::size_t>(job->slot)].reset();
                 m_runningCount = 0;
             }
         }
@@ -721,7 +723,8 @@ class ScreenshotOcrRecognitionService::Impl final {
                      {{QStringLiteral("child_pid"), pid},
                       {QStringLiteral("slot_count"), static_cast<int>(m_slots.size())},
                       {QStringLiteral("shared_memory_bytes"),
-                       static_cast<qint64>(m_slotBytes * m_slots.size())}});
+                       static_cast<qint64>(static_cast<std::size_t>(m_slotBytes) *
+                                           m_slots.size())}});
              },
              [this, valid](QByteArray bytes) {
                  if (valid())
@@ -805,7 +808,7 @@ class ScreenshotOcrRecognitionService::Impl final {
 
     int acquireSlot() const {
         for (int index = 0; index < static_cast<int>(m_slots.size()); ++index)
-            if (!m_slots[index])
+            if (!m_slots[static_cast<std::size_t>(index)])
                 return index;
         return -1;
     }
@@ -928,8 +931,8 @@ class ScreenshotOcrRecognitionService::Impl final {
                 }
                 return;
             }
-            const quint64 sequence = ++m_slotSequences[slot];
-            m_slots[slot] = job;
+            const quint64 sequence = ++m_slotSequences[static_cast<std::size_t>(slot)];
+            m_slots[static_cast<std::size_t>(slot)] = job;
             job->slot = slot;
             job->running = true;
             job->processSubmitted = true;
@@ -972,12 +975,12 @@ class ScreenshotOcrRecognitionService::Impl final {
                 (static_cast<quint32>(static_cast<quint8>(m_readBuffer.at(1))) << 8) |
                 (static_cast<quint32>(static_cast<quint8>(m_readBuffer.at(2))) << 16) |
                 (static_cast<quint32>(static_cast<quint8>(m_readBuffer.at(3))) << 24);
-            const quint16 version =
+            const quint16 version = static_cast<quint16>(
                 static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(4))) |
-                (static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(5))) << 8);
-            const quint16 kind =
+                (static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(5))) << 8));
+            const quint16 kind = static_cast<quint16>(
                 static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(6))) |
-                (static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(7))) << 8);
+                (static_cast<quint16>(static_cast<quint8>(m_readBuffer.at(7))) << 8));
             qsizetype offset = 8;
             quint64 id = 0;
             quint32 length = 0;
@@ -1084,7 +1087,7 @@ class ScreenshotOcrRecognitionService::Impl final {
                         }
                     ScreenshotOcrLine line;
                     line.text = std::move(text);
-                    line.confidence = confidence;
+                    line.confidence = static_cast<qreal>(confidence);
                     line.quad =
                         quadFromValues(points, job->request.canvasRect, job->request.image.size());
                     line.direction = textDirectionForQuad(line.quad);
@@ -1096,7 +1099,7 @@ class ScreenshotOcrRecognitionService::Impl final {
                                                            "Text recognition failed");
             }
             if (job->slot >= 0) {
-                m_slots[job->slot].reset();
+                m_slots[static_cast<std::size_t>(job->slot)].reset();
             }
             job->slot = -1;
             job->running = false;
@@ -1192,7 +1195,7 @@ class ScreenshotOcrRecognitionService::Impl final {
             QStringLiteral("ocr.failed"), error, fields);
         m_pending.erase(std::remove(m_pending.begin(), m_pending.end(), job), m_pending.end());
         if (job->slot >= 0) {
-            m_slots[job->slot].reset();
+            m_slots[static_cast<std::size_t>(job->slot)].reset();
             job->slot = -1;
             job->running = false;
             job->processSubmitted = false;
